agp-gemini-cli/pyproject.toml



[project]
name = "agp-gemini-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "alejandro",email = "noemisanalex@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "typer (>=0.17.4,<0.18.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "speechrecognition (>=3.14.3,<4.0.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "structlog (>=25.4.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "mypy (>=1.10.0,<2.0.0)"
]
